requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
argcomplete==1.10.0
attrs==25.1.0
beautifulsoup4==4.8.0
bidict==0.23.1
bs4==0.0.2
certifi==2025.1.31
chardet==3.0.4
charset-normalizer==3.4.1
click==8.1.8
cmudict==1.0.32
colorama==0.4.6
dataclasses-json==0.6.7
Distance==0.1.3
docx2txt==0.8
easyocr==1.7.2
EbookLib==0.17.1
emoji==1.2.0
et_xmlfile==2.0.0
extract-msg==0.23.1
faiss-cpu==1.10.0
fastapi==0.115.8
filelock==3.13.1
frozenlist==1.5.0
fsspec==2024.6.1
greenlet==3.1.1
h11==0.14.0
hangul-jamo==1.0.1
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.29.2
idna==3.10
imageio==2.37.0
IMAPClient==2.1.0
importlib_metadata==8.6.1
importlib_resources==6.5.2
iniconfig==2.0.0
ipaddress==1.0.23
jamo==0.4.1
Jinja2==3.1.4
joblib==1.4.2
jsonlines==4.0.0
jsonpatch==1.33
jsonpointer==3.0.0
kiwipiepy==0.20.3
kiwipiepy_model==0.20.0
kollocate==0.0.2
koparadigm==0.10.0
langchain==0.3.17
langchain-community==0.3.16
langchain-core==0.3.33
langchain-huggingface==0.1.2
langchain-text-splitters==0.3.5
langsmith==0.3.4
lazy_loader==0.4
lxml==5.3.0
MarkupSafe==2.1.5
marshmallow==3.26.0
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
natsort==8.4.0
networkx==3.3
ninja==1.11.1.3
numpy==1.26.4
olefile==0.46
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
orjson==3.10.15
packaging==24.2
pandas==2.2.3
pdfminer.six==20181108
pecab==1.0.8
pillow==11.0.0
pluggy==1.5.0
propcache==0.2.1
psutil==7.0.0
pyarrow==19.0.0
pyclipper==1.3.0.post6
pycryptodome==3.22.0
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
PyPDF2==3.0.1
pytest==8.3.4
python-bidi==0.6.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
python-pptx==0.6.18
pytz==2025.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
safetensors==0.5.3
scikit-image==0.25.1
scikit-learn==1.6.1
scipy==1.15.1
sentence-transformers==3.4.1
shapely==2.0.7
six==1.12.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SpeechRecognition==3.8.1
SQLAlchemy==2.0.37
starlette==0.45.3
sympy==1.13.1
tabulate
tenacity==9.0.0
textract==1.6.3
tokenizers
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tossi==0.3.1
tqdm==4.67.1
transformers==4.48.2
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==1.5.1
Unidecode==1.3.8
urllib3==2.3.0
uvicorn==0.34.0
watchdog==6.0.0
Whoosh==2.7.4
xlrd==1.2.0
XlsxWriter==3.2.2
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0